main.py
import os
import json
import logging
from datetime import datetime
from typing import List, Dict
from googleapiclient.discovery import build
from google.oauth2.credentials import Credentials
from transformers import AutoTokenizer, AutoModelForSequenceClassification
import spacy
from langchain_core.tools import tool
from langchain_community.vectorstores import FAISS
from langchain_openai import OpenAIEmbeddings, ChatOpenAI

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class EnhancedGmailAPIModule:
    def __init__(self, credentials_path: str):
        self.credentials = self._authenticate(credentials_path)
        self.service = build('gmail', 'v1', credentials=self.credentials)
        self.rate_limit = 250  # Requests per minute [7]
        
    def _authenticate(self, path: str) -> Credentials:
        """OAuth 2.0 with token refresh handling"""
        return Credentials.from_authorized_user_file(path)
    
    @tool
    def get_emails(self, query: str = "is:inbox is:unread") -> List[Dict]:
        """Fetch emails with advanced query capabilities"""
        try:
            result = self.service.users().messages().list(
                userId='me',
                q=query,
                maxResults=20
            ).execute()
            return result.get('messages', [])
        except Exception as e:
            logger.error(f"API Error: {str(e)}")
            return []
    
    @tool
    def send_email(self, to: str, subject: str, body: str) -> bool:
        """Send email with structured content validation"""
        message = {
            'raw': f"To: {to}\nSubject: {subject}\n\n{body}"
        }
        try:
            self.service.users().messages().send(
                userId='me',
                body=message
            ).execute()
            return True
        except Exception as e:
            logger.error(f"Send failed: {str(e)}")
            return False

class RAGEnhancedUnderstanding:
    def __init__(self):
        self.nlp = spacy.load("en_core_web_lg")
        self.tokenizer = AutoTokenizer.from_pretrained("bert-base-uncased")
        self.intent_model = AutoModelForSequenceClassification.from_pretrained("email-intent-bert")
        self.vector_store = FAISS.load_local("email_index", OpenAIEmbeddings())  # [5][11]
        self.llm = ChatOpenAI(model="gpt-4-1106-preview")  # [3][6]
        
    def analyze_email(self, text: str) -> Dict:
        """Multi-stage analysis with RAG augmentation"""
        # Intent classification
        inputs = self.tokenizer(text, return_tensors="pt")
        intent_logits = self.intent_model(**inputs).logits
        intent = self._decode_intent(intent_logits)
        
        # Entity extraction
        doc = self.nlp(text)
        entities = [(ent.text, ent.label_) for ent in doc.ents]
        
        # RAG context retrieval
        relevant_docs = self.vector_store.similarity_search(text, k=3)
        context = "\n".join([d.page_content for d in relevant_docs])
        
        # LLM-enhanced analysis
        analysis_prompt = f"""
        Analyze email with context:
        Email: {text}
        Context: {context}
        Output JSON with: summary, urgency_score (1-5), required_actions
        """
        analysis = self.llm.invoke(analysis_prompt).content
        return json.loads(analysis)

class ActionOrchestrator:
    def __init__(self):
        self.rules = self._load_rules()
        self.llm = ChatOpenAI(model="gpt-4-1106-preview")
        
    def determine_actions(self, analysis: Dict) -> List[str]:
        """Hybrid rule-based/LLM decision making"""
        # Rule-based priority
        if analysis['urgency_score'] >= 4:
            return ["notify_user", "create_task"]
            
        # LLM-generated actions
        prompt = f"""
        Given analysis: {json.dumps(analysis)}
        Generate appropriate actions from: reply, archive, forward, task
        """
        actions = self.llm.invoke(prompt).content
        return json.loads(actions)
        
    def _load_rules(self) -> Dict:
        """Load business rules from YAML"""
        with open("rules.yaml") as f:
            return yaml.safe_load(f)

class LearningModule:
    def __init__(self):
        self.feedback_db = SQLiteDatabase("feedback.db")
        self.retraining_interval = 86400  # 24 hours
        
    def log_feedback(self, email_id: str, accepted: bool):
        """Store explicit/implicit feedback"""
        self.feedback_db.execute(
            "INSERT INTO feedback VALUES (?, ?, ?)",
            (email_id, accepted, datetime.now())
        )
        
    def retrain_model(self):
        """Periodic model updating"""
        if time.time() - self.last_retrain > self.retraining_interval:
            # Implement online learning logic [9]
            pass

class EmailAutomationAgent:
    def __init__(self):
        self.gmail = EnhancedGmailAPIModule("credentials.json")
        self.analyzer = RAGEnhancedUnderstanding()
        self.orchestrator = ActionOrchestrator()
        self.learning = LearningModule()
        
    def process_emails(self):
        """Main processing loop with error handling"""
        try:
            emails = self.gmail.get_emails()
            for email in emails:
                content = self._get_email_content(email['id'])
                analysis = self.analyzer.analyze_email(content)
                actions = self.orchestrator.determine_actions(analysis)
                self._execute_actions(email['id'], actions)
        except Exception as e:
            logger.error(f"Processing failed: {str(e)}")
            # Implement circuit breaker [1][7]

if __name__ == "__main__":
    agent = EmailAutomationAgent()
    agent.process_emails()
