requirement.txt
# Core Dependencies
python==3.10.12
google-api-python-client==2.108.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0

# NLP & Machine Learning
spacy==3.7.4
scikit-learn==1.4.1.post1
transformers==4.38.2
sentence-transformers==2.3.1
torch==2.2.0
numpy==1.26.4
pandas==2.2.0

# LangChain & LLM Integration
langchain-core==0.1.33
langchain-community==0.0.29
langchain-openai==0.0.8
openai==1.12.0
tiktoken==0.6.0

# Vector Stores & Search
faiss-cpu==1.8.0
sentencepiece==0.2.0

# Task Queues & Async Processing
celery==5.3.6
redis==5.0.3

# Additional Utilities
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
tqdm==4.66.2
loguru==0.7.2

# Production Dependencies
gunicorn==21.2.0
uvicorn==0.27.0
httpx==0.27.0
pytest==8.0.2
